PAE-1384: rename is-report-complete to assert-report-complete
