+++ src/reports/application/assert-report-complete.js
