+++ src/reports/routes/status.js
@@ -3,7 +3,7 @@ import Joi from 'joi'
 import { StatusCodes } from 'http-status-codes'
 
 import { auditReportStatusTransition } from '#reports/application/audit.js'
-import { assertReportComplete } from '#reports/application/is-report-complete.js'
+import { assertReportComplete } from '#reports/application/assert-report-complete.js'
 import { fetchCurrentReport } from '#reports/application/report-service.js'
 import { getOperatorCategory } from '#reports/domain/operator-category.js'
 import { REPORT_STATUS } from '#reports/domain/report-status.js'
